[#7] Rename resetpassword package
