+++ lib/navigator.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:injectable/injectable.dart';
 import 'package:survey/page/home/home_page.dart';
 import 'package:survey/page/login/login_page.dart';
-import 'package:survey/page/reset_password/reset_password_page.dart';
+import 'package:survey/page/resetpassword/reset_password_page.dart';
 
 const String _routeLogin = 'login';
 const String _routeHome = 'home';
